+++ CHANGES.md
@@ -1,5 +1,7 @@
 # UNRELEASED
 
+- Fix `FlexData.Elpi.make` when called with a name after compilation is over
+- Fix `RawQuery.mk_Arg` can only be called at compile time
 - Fix printing of clauses in trace
 
 # v1.16.5 (July 2022)